refactor auth.js into separate files
