+++ src/models/auth/AuthorizationCode.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose')
+
+const authorizationCodeSchema = mongoose.Schema(
+  {
+    authorizationCode: String,
+    scope: String,
+    expiresAt: Date,
+    redirectUri: String,
+    client: { type: mongoose.Schema.Types.ObjectId, ref: 'Client' },
+    user: { type: mongoose.Schema.Types.ObjectId, ref: 'UserTests' }
+  }
+
+)
+
+module.exports = mongoose.model('AuthorizationCodeSchema', authorizationCodeSchema)
